[REFACTOR]  APIManager과 extension으로 리팩토링
