+++ NetWorkProject/BeerAPI/BeerListViewController.swift
@@ -1,0 +1,29 @@
+//
+//  BeerListViewController.swift
+//  NetWorkProject
+//
+//  Created by 남현정 on 2024/01/17.
+//
+
+import UIKit
+
+class BeerListViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
