+++ NetWorkProject/BeerAPI/BeerRandomViewController.swift
@@ -1,0 +1,17 @@
+//
+//  BeerRandomViewController.swift
+//  NetWorkProject
+//
+//  Created by 남현정 on 2024/01/17.
+//
+
+import UIKit
+
+// 두 번 째 탭
+class BeerRandomViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+}
